server/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
Django==4.2.9
djangorestframework==3.14.0
easyocr==1.7.1
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.41
idna==3.6
imageio==2.33.1
importlib-resources==6.1.1
Jinja2==3.1.3
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.4
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.3
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.2.0
pillow==10.2.0
psutil==5.9.8
py-cpuinfo==9.0.0
pyclipper==1.3.0.post5
pyparsing==3.1.1
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.22.0
scipy==1.12.0
seaborn==0.13.1
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sqlparse==0.4.4
sympy==1.12
thop==0.1.1.post2209072238
tifffile==2023.12.9
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
ultralytics==8.1.5
urllib3==2.1.0
zipp==3.17.0
